requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
fastapi==0.115.11
idna==3.10
iniconfig==2.0.0
numpy==2.2.3
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
requests==2.32.3
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
